pyproject.toml
[project]
name = "pic-cmp"
version = "0.2.0"
description = "Copyright detection system using SAM3 and RoMaV2"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]",
    "python-multipart",
    "pillow",
    "requests",
    "openpyxl",
]

[project.optional-dependencies]
test = [
    "pytest>=7",
    "httpx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "backend"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
